# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Delhi Ops Dashboard API
  version: 1.0.0
  description: Real-time urban monitoring dashboard for Delhi - Air Quality, Crime, Traffic, Cameras, and Incidents.
servers:
  - url: /api/v1
    description: API v1
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 500
        default: 50
    sortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    sortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: desc
    fromDateParam:
      name: from_date
      in: query
      schema:
        type: string
        format: date-time
    toDateParam:
      name: to_date
      in: query
      schema:
        type: string
        format: date-time
  schemas:
    Location:
      type: object
      required: [lat, lng]
      properties:
        lat:
          type: number
          minimum: -90
          maximum: 90
        lng:
          type: number
          minimum: -180
          maximum: 180
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data: {}
    PaginatedResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data: {}
        pagination:
          $ref: '#/components/schemas/Pagination'
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details: {}
            requestId:
              type: string
...
    TrafficData:
      type: object
      properties:
        id:
          type: integer
        congestion_level:
          type: number
          minimum: 0
          maximum: 100
        speed:
          type: number
          minimum: 0
        timestamp:
          type: string
          format: date-time
    CreateTrafficRequest:
      type: object
      required: [congestion_level, speed]
      properties:
        congestion_level:
          type: number
          minimum: 0
          maximum: 100
        speed:
          type: number
          minimum: 0
        timestamp:
          type: string
          format: date-time
...
    Incident:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
        severity:
          type: string
        timestamp:
          type: string
          format: date-time
        description:
          type: string
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
        location:
          $ref: '#/components/schemas/Location'
    CreateIncidentRequest:
      type: object
      required: [type, location]
      properties:
        type:
          type: string
          maxLength: 100
        severity:
          type: string
          maxLength: 50
        location:
          $ref: '#/components/schemas/Location'
        timestamp:
          type: string
          format: date-time
        description:
          type: string
          maxLength: 5000
        status:
          type: string
          enum: [open, in_progress, resolved, closed]
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequest:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
security:
  - bearerAuth: []
paths:
...
  /traffic:
    get:
      tags: [Traffic]
      summary: List traffic data
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/sortByParam'
        - $ref: '#/components/parameters/sortOrderParam'
        - $ref: '#/components/parameters/fromDateParam'
        - $ref: '#/components/parameters/toDateParam'
        - name: congestion_min
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
        - name: congestion_max
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
      responses:
        '200':
          description: Paginated traffic data
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/TrafficData'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Traffic]
      summary: Create traffic data record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTrafficRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/TrafficData'
        '400':
          $ref: '#/components/responses/BadRequest'
...
  /incidents:
    get:
      tags: [Incidents]
      summary: List incidents
      parameters:
        - $ref: '#/components/parameters/pageParam'
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/sortByParam'
        - $ref: '#/components/parameters/sortOrderParam'
        - $ref: '#/components/parameters/fromDateParam'
        - $ref: '#/components/parameters/toDateParam'
        - name: type
          in: query
          schema:
            type: string
            maxLength: 100
        - name: severity
          in: query
          schema:
            type: string
            maxLength: 50
        - name: status
          in: query
          schema:
            type: string
            enum: [open, in_progress, resolved, closed]
      responses:
        '200':
          description: Paginated incidents
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PaginatedResponse'
                  - properties:
                      data:
                        type: array
                        items:
                          $ref: '#/components/schemas/Incident'
        '401':
          $ref: '#/components/responses/Unauthorized'
    post:
      tags: [Incidents]
      summary: Create incident
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateIncidentRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - properties:
                      data:
                        $ref: '#/components/schemas/Incident'
        '400':
          $ref: '#/components/responses/BadRequest'
...